# lib/api-zod/src/generated/types/geminiConversation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Betting Analyst Toolkit API
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: Date;
}