# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Betting Analyst Toolkit API
 * OpenAPI spec version: 0.1.0
 */

export interface Prediction {
  id: number;
  sport: string;
  homeTeam: string;
  awayTeam: string;
  matchDate: Date;
  predictedOutcome: string;
  confidence: number;
  odds: number;
  kellyFraction: number;
  reasoning: string;
  status: string;
  /** @nullable */
  actualOutcome: string | null;
  sentToTelegram: boolean;
  createdAt: Date;
  updatedAt: Date;
}