# lib/api-zod/src/generated/types/predictionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Betting Analyst Toolkit API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionInput {
  sport: string;
  homeTeam: string;
  awayTeam: string;
  matchDate: Date;
  homeOdds: number;
  /** @nullable */
  drawOdds?: number | null;
  awayOdds: number;
}