# lib/api-zod/src/generated/types/predictionResultInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Betting Analyst Toolkit API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionResultInput {
  actualOutcome: string;
  status: string;
}